feat: add `tabby scheduler` command
