+++ Cargo.toml
@@ -19,3 +19,5 @@ serde = { version = "1.0", features = ["derive"] }
 serdeconv = "0.4.1"
 tokio = "1.28"
 tokio-util = "0.7"
+tracing = "0.1"
+tracing-subscriber = "0.3"
